src/main/java/guru/springframework/myreceipe/services/RecipeServiceImpl.java
package guru.springframework.myreceipe.services;

import guru.springframework.myreceipe.domain.Recipe;
import guru.springframework.myreceipe.repositories.RecipeRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;

import java.util.HashSet;
import java.util.Set;

@RequiredArgsConstructor
@Service
public class RecipeServiceImpl implements RecipeService {
    private final RecipeRepository recipeRepository;

    @Override
    public Set<Recipe> getRecipes() {
        Set<Recipe> recipeSet = new HashSet<>();
        System.out.println(" it is in service");
        recipeRepository.findAll().iterator().forEachRemaining(recipeSet::add);
        return recipeSet;
    }
}
